refactor: move myscraper module to repo root
