--- go/myscraper/go.mod
@@ -1,3 +1,0 @@
-module github.com/azuki774/myscrapers/go/myscraper
-
-go 1.24.0
