+++ myscraper/go.mod
@@ -1,0 +1,3 @@
+module github.com/azuki774/myscrapers/myscraper
+
+go 1.24.0
