+++ myscraper/scripts/dev-shell-smoke.sh
